requirements/dev.txt
fastapi==0.104.1
fastapi-cache2==0.2.0
SQLAlchemy==2.0.23
protobuf==4.25.1
grpcio==1.60.0
grpcio-tools==1.60.0
python-dotenv==1.0.0
psycopg2==2.9.9
asyncpg==0.29.0
aiohttp==3.9.1